+++ touki/Framework/Polyfills/System/ArgumentOutOfRangeExtensions.cs
@@ -1,4 +1,4 @@
-﻿// Copyright (c) 2025 Jeremy W Kuhne
+// Copyright (c) 2025 Jeremy W Kuhne
 // SPDX-License-Identifier: MIT
 // See LICENSE file in the project root for full license information
 
@@ -7,7 +7,9 @@
 
 using Touki.Framework.Resources;
 
-namespace Touki.Exceptions;
+using Touki;
+
+namespace System;
 
 /// <summary>
 ///  Helper to allow using new patterns for throwing <see cref="ArgumentOutOfRangeException"/>s.
