+++ touki/Framework/Polyfills/System/ConvertExtensions.cs
