+++ touki/Framework/Polyfills/System/HashCodeExtensions.cs
