+++ touki/Framework/Polyfills/System/HexConverter.cs
