+++ touki/Framework/Polyfills/System/ISpanFormattable.cs
