+++ touki/Framework/Polyfills/System/MemoryExtensions.TryWrite.cs
@@ -4,7 +4,9 @@
 
 // Adapted from dotnet/runtime MemoryExtensions.TryWriteInterpolatedStringHandler.cs (MIT licensed).
 
-namespace Touki;
+using Touki;
+
+namespace System;
 
 public static partial class MemoryExtensions
 {
