+++ touki/Framework/Polyfills/System/Number.BigInteger.cs
