+++ touki/Framework/Polyfills/System/Number.DiyFp.cs
