+++ touki/Framework/Polyfills/System/Number.Dragon4Double.cs
