+++ touki/Framework/Polyfills/System/Number.Formatting.cs
