+++ touki/Framework/Polyfills/System/Number.NumberBuffer.cs
