+++ touki/Framework/Polyfills/System/Number.NumberToFloatingPointBits.cs
