+++ touki/Framework/Polyfills/System/ObjectDisposedExtensions.cs
@@ -1,8 +1,10 @@
-﻿// Copyright (c) 2025 Jeremy W Kuhne
+// Copyright (c) 2025 Jeremy W Kuhne
 // SPDX-License-Identifier: MIT
 // See LICENSE file in the project root for full license information
 
-namespace Touki.Exceptions;
+using Touki;
+
+namespace System;
 
 /// <summary>
 ///  ObjectDisposedException helper to allow using new patterns for throwing <see cref="ObjectDisposedException"/>s.
