+++ touki/Framework/Polyfills/System/RandomExtensions.cs
