+++ touki/Framework/Polyfills/System/SpanExtensions.InRange.cs
@@ -2,7 +2,9 @@
 // SPDX-License-Identifier: MIT
 // See LICENSE file in the project root for full license information
 
-namespace Touki;
+using Touki;
+
+namespace System;
 
 public static partial class SpanExtensions
 {
