+++ touki/Framework/Polyfills/System/SpanExtensions.SpanSplitEnumerator.cs
@@ -1,4 +1,4 @@
-﻿// Copyright (c) 2025 Jeremy W Kuhne
+// Copyright (c) 2025 Jeremy W Kuhne
 // SPDX-License-Identifier: MIT
 // See LICENSE file in the project root for full license information
 
@@ -9,7 +9,9 @@
 
 using System.Collections;
 
-namespace Touki;
+using Touki;
+
+namespace System;
 
 public static partial class SpanExtensions
 {
