+++ touki/Framework/Polyfills/System/StringExtensions.cs
